Move component props to separate files
